# openapi/altus4.openapi.yaml
openapi: 3.0.3
info:
  title: Altus4 Core API
  version: 0.4.0
  description: |
    Complete API specification for Altus4 - AI-Enhanced MySQL Full-Text Search Engine.

    ## Authentication

    Altus4 uses dual authentication:
    - **JWT Bearer Tokens**: For user management, API key management, database management, and analytics
    - **API Key Bearer Tokens**: For search operations and B2B service integration

    API Key format: `altus4_sk_{env}_{secret}` where env is 'test' or 'live'.
servers:
  - url: http://localhost:3000
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for user authentication
    apiKeyAuth:
      type: http
      scheme: bearer
      description: API key for search operations (format altus4_sk_{env}_{secret})
  schemas:
    ApiResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        data:
          description: Response data (varies by endpoint)
        error:
          type: object
          properties:
            code:
              type: string
              description: Error code for programmatic handling
            message:
              type: string
              description: Human-readable error message
            details:
              description: Additional error details
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
            version:
              type: string
            executionTime:
              type: number
              description: Request execution time in milliseconds
            apiKeyTier:
              type: string
              description: Present on API key authenticated responses
...
    DatabaseConnection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        host:
          type: string
        port:
          type: integer
        database:
          type: string
        username:
          type: string
        ssl:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Bad Request - Validation error
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: VALIDATION_ERROR
                      message:
                        type: string
                        example: Invalid request parameters
    Unauthorized:
      description: Unauthorized - Missing or invalid token
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: INVALID_API_KEY
                      message:
                        type: string
                        example: Missing or invalid API key
...
    NotFound:
      description: Not Found - Resource not found
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: NOT_FOUND
                      message:
                        type: string
                        example: Resource not found
...
paths:
...
  /api/v1/databases:
    get:
      tags: [Database]
      summary: List database connections
      description: Get all database connections for the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Database connections retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/DatabaseConnection'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Database]
      summary: Add database connection
      description: Add a new database connection for the authenticated user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, host, database, username]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                host:
                  type: string
                  minLength: 1
                port:
                  type: integer
                  minimum: 1
                  maximum: 65535
                  default: 3306
                database:
                  type: string
                  minLength: 1
                username:
                  type: string
                  minLength: 1
                password:
                  type: string
                  description: Database password (will be encrypted)
                ssl:
                  type: boolean
                  default: false
      responses:
        '201':
          description: Database connection added successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DatabaseConnection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/databases/status:
    get:
      tags: [Database]
      summary: Get connection statuses
      description: Get connection status for all user databases
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Connection statuses retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        additionalProperties:
                          type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/databases/{connectionId}:
    parameters:
      - in: path
        name: connectionId
        required: true
        schema:
          type: string
          format: uuid
        description: Database connection ID

    get:
      tags: [Database]
      summary: Get database connection
      description: Get details of a specific database connection
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Database connection retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DatabaseConnection'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Database]
      summary: Update database connection
      description: Update an existing database connection
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 255
                host:
                  type: string
                  minLength: 1
                port:
                  type: integer
                  minimum: 1
                  maximum: 65535
                database:
                  type: string
                  minLength: 1
                username:
                  type: string
                  minLength: 1
                password:
                  type: string
                  description: Database password (will be encrypted)
                ssl:
                  type: boolean
      responses:
        '200':
          description: Database connection updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/DatabaseConnection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Database]
      summary: Remove database connection
      description: Remove a database connection
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Database connection removed successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          success:
                            type: boolean
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/databases/{connectionId}/test:
    parameters:
      - in: path
        name: connectionId
        required: true
        schema:
          type: string
          format: uuid
        description: Database connection ID

    post:
      tags: [Database]
      summary: Test database connection
      description: Test connectivity to a database
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Connection test completed
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          connected:
                            type: boolean
                          message:
                            type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/databases/{connectionId}/schema:
    parameters:
      - in: path
        name: connectionId
        required: true
        schema:
          type: string
          format: uuid
        description: Database connection ID

    get:
      tags: [Database]
      summary: Get database schema
      description: Discover and return the schema of a connected database
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Database schema retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          type: object
                          properties:
                            database:
                              type: string
                            table:
                              type: string
                            columns:
                              type: array
                              items:
                                type: object
                                properties:
                                  name:
                                    type: string
                                  type:
                                    type: string
                                  isFullTextIndexed:
                                    type: boolean
                                  isSearchable:
                                    type: boolean
                                  dataPreview:
                                    type: array
                                    items:
                                      type: string
                            fullTextIndexes:
                              type: array
                              items:
                                type: object
                                properties:
                                  name:
                                    type: string
                                  columns:
                                    type: array
                                    items:
                                      type: string
                                  type:
                                    type: string
                                    enum: [FULLTEXT]
                                  cardinality:
                                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # Search Endpoints
...